requirements.txt
#tensorflow==2.15.0
ultralytics==8.0.237
python-dotenv==1.0.0
tqdm
azure-storage-blob==12.19.0
azure-ai-vision==0.15.1b1
azure-search-documents
pandas
numpy
annoy
opencv-python==4.9.0.80
streamlit==1.29.0
st_clickable_images
streamlit_selectable_image_gallery
langchain==0.1.7
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.38
langchain-openai==0.1.1